+++ tests/test_tablib_dbfpy_packages_utils.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python
-"""Tests for tablib.packages.dbfpy."""
+"""Tests for tablib._vendor.dbfpy."""
 
 import datetime as dt
 import unittest
 
-from tablib.packages.dbfpy import utils
+from tablib._vendor.dbfpy import utils
 
 
 class UtilsUnzfillTestCase(unittest.TestCase):
